requirements.txt
Django==5.2
django-extensions==4.1
ipython==8.35.0
django-debug-toolbar==5.2.0
pillow==11.2.1
python-dotenv==1.1.0
social-auth-app-django==5.4.3
Werkzeug==3.1.3
pyOpenSSL==25.1.0
psycopg2==2.9.10
django-simple-captcha==0.6.2
redis==6.2.0
hiredis==3.2.1
